backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.6.0
gunicorn==23.0.0

